May 30, 2026: Now Admin can not add more than 50 Medicine records in Database and cannot delete a Medicine record if less than or equal to 20

+++ src/controllers/admin/manage/medicines/delete.controller.js
@@ -4,9 +4,22 @@ import path from "path";
 
 const DeleteMedicineRecordCtrl = async ( req, res, next ) => {
     
+    
+    const totalMedicineRecordsInDb = await MedicinesCollection.countDocuments();
+    
+    /*
+    * If database contains less than or equal to 20 records, than 
+    * records cannot be deleted.
+    */
+   if ( totalMedicineRecordsInDb <= 20 ) {
+       return res.redirect( 
+           `/admin/manage/medicines/?message=This Demo Project must contains at least 20 records. Add more records to delete` 
+        );
+    }
+
     const deletedMedicineRecordFromDB = 
     await MedicinesCollection.findByIdAndDelete( req.params.id );
-
+    
     fs.unlink( 
         path.join(
             process.cwd(), "public", "uploads", "medicines", deletedMedicineRecordFromDB.medicineImage
